Extract removeDiffSign feature to its own file
