+++ src/content.js
@@ -23,6 +23,7 @@ import autoLoadMoreNews from './libs/auto-load-more-news';
 import addOPLabels from './libs/op-labels';
 import addReleasesTab from './libs/add-releases-tab';
 import scrollToTopOnCollapse from './libs/scroll-to-top-on-collapse';
+import removeDiffSigns from './libs/remove-diff-signs';
 
 import * as icons from './libs/icons';
 import * as pageDetect from './libs/page-detect';
@@ -292,41 +293,6 @@ function addPatchDiffLinks() {
 			<a href={`${commitUrl}.diff`} class="sha">diff</a>
 		</span>
 	);
-}
-
-function removeSelectableWhiteSpaceFromDiffs() {
-	for (const commentBtn of select.all('.add-line-comment')) {
-		for (const node of commentBtn.childNodes) {
-			if (node.nodeType === Node.TEXT_NODE) {
-				node.remove();
-			}
-		}
-	}
-}
-
-/* Lasciate ogne speranza, voi ch'intrate. */
-function removeDiffSigns() {
-	for (const line of select.all('.diff-table tr:not(.refined-github-diff-signs)')) {
-		line.classList.add('refined-github-diff-signs');
-		for (const code of select.all('.blob-code-inner', line)) {
-			// Drop -, + or space
-			code.firstChild.textContent = code.firstChild.textContent.slice(1);
-
-			// If a line is empty, the next line will collapse
-			if (code.textContent.length === 0) {
-				code.prepend(' ');
-			}
-		}
-	}
-}
-
-function removeDiffSignsAndWatchExpansions() {
-	removeSelectableWhiteSpaceFromDiffs();
-	removeDiffSigns();
-	for (const file of $('.diff-table:not(.rgh-watching-lines)').has('.diff-expander')) {
-		file.classList.add('rgh-watching-lines');
-		observeEl(file.tBodies[0], removeDiffSigns);
-	}
 }
 
 function markMergeCommitsInList() {
@@ -671,13 +637,7 @@ async function onDomReady() {
 			safely(addReadmeButtons);
 			safely(addDiffViewWithoutWhitespaceOption);
 			safely(enableCopyOnY.destroy);
-
-			safely(() => {
-				const diffElements = select('.js-discussion, #files');
-				if (diffElements) {
-					observeEl(diffElements, removeDiffSignsAndWatchExpansions, {childList: true, subtree: true});
-				}
-			});
+			safely(removeDiffSigns);
 
 			safely(() => {
 				for (const a of select.all('a[href]')) {
